# docs/swagger.yaml
definitions:
...
  controllers.PostCreateItemRequest:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: number
      qty:
        type: integer
    required:
    - description
    - name
    - price
    - qty
    type: object
  controllers.PostCreateOrderRequest:
    properties:
      address:
        type: string
      phone_number:
        type: string
      voucher_code:
        type: string
    type: object
...
  controllers.PutEditItemRequest:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: number
      qty:
        type: integer
    required:
    - description
    - name
    - price
    - qty
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a API for Online Shopping Cart API.
  title: Online Shopping Cart API
  version: "1.0"
paths:
...
  /item:
    post:
      description: add item for admin user, need credential ADMIN user only
      parameters:
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6ImFkbWluQGV4YW1wbGUuY29tIiwiZW1haWwiOiJhZG1pbkBleGFtcGxlLmNvbSIsInR5cGUiOiJBRE1JTiIsImV4cCI6MTcyNTU5ODc3OX0.JamULnKlo3q38ZgIhfBOUI8U2WEv4nNfaLYvodtIx0c
        description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body Request
        in: body
        name: tags
        required: true
        schema:
          $ref: '#/definitions/controllers.PostCreateItemRequest'
      produces:
      - application/json
      responses: {}
      summary: add item for admin user
      tags:
      - Item
  /item/{id}:
    delete:
      description: delete item for admin user, need credential ADMIN user only
      parameters:
      - description: delete by id
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6ImFkbWluQGV4YW1wbGUuY29tIiwiZW1haWwiOiJhZG1pbkBleGFtcGxlLmNvbSIsInR5cGUiOiJBRE1JTiIsImV4cCI6MTcyNTU5ODc3OX0.JamULnKlo3q38ZgIhfBOUI8U2WEv4nNfaLYvodtIx0c
        description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: delete item for admin user
      tags:
      - Item
    get:
      description: get one item detail, need credentials
      parameters:
      - description: get detail by id
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer {token}
        description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: get one item detail
      tags:
      - Item
    put:
      description: edit item for admin user, need credential ADMIN user only
      parameters:
      - description: edit by id
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6ImFkbWluQGV4YW1wbGUuY29tIiwiZW1haWwiOiJhZG1pbkBleGFtcGxlLmNvbSIsInR5cGUiOiJBRE1JTiIsImV4cCI6MTcyNTU5ODc3OX0.JamULnKlo3q38ZgIhfBOUI8U2WEv4nNfaLYvodtIx0c
        description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body Request
        in: body
        name: tags
        required: true
        schema:
          $ref: '#/definitions/controllers.PutEditItemRequest'
      produces:
      - application/json
      responses: {}
      summary: edit item for admin user
      tags:
      - Item
  /item/all:
    get:
      description: get all item need credentials
      parameters:
      - default: Bearer {token}
        description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: name item
        in: query
        name: name
        type: string
      - description: min price item
        in: query
        name: min_price
        type: string
      - description: max price item
        in: query
        name: max_price
        type: string
      produces:
      - application/json
      responses: {}
      summary: get all item
      tags:
      - Item
...
  /order:
    post:
      description: create order from cart, need credential
      parameters:
      - default: Bearer {token}
        description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body Request
        in: body
        name: tags
        required: true
        schema:
          $ref: '#/definitions/controllers.PostCreateOrderRequest'
      produces:
      - application/json
      responses: {}
      summary: create order from cart
      tags:
      - Order
  /order/{id}:
    get:
      description: get one own order detail, need credentials
      parameters:
      - description: get detail by id
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer {token}
        description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: get one own order detail
      tags:
      - Order
  /order/all:
    get:
      description: get all own order need credentials
      parameters:
      - default: Bearer {token}
        description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: get all own order
      tags:
      - Order
...
swagger: "2.0"